Разбиение main.js на модули
